+++ Thomas DAQ Work/Teensy/AD5764Lib/AD5764.h
@@ -22,7 +22,7 @@ class AD5764 {
 
   //set voltage output for one of the 4 DAC channels (max output: ±10V)
   //vout: expecting pre-converted value (0-65535 Binary encoding), see AD5764 documentation.
-  void SetDataRegister(uint16_t vout, int dac_channel);
+  void SetDataRegister(uint16_t vout, uint8_t dac_channel);
 
   private:
   int _cs, _ldac, _clr;
